POC ONLY - live preview updates.
This POC explores what it would look like to have live updates of the
preview panel as content is updated in the CMS.
The general approach is:

- When the CMSPreview get variable is present and the current user has
  CMS access, the template rendered wraps rendered properties with some
sort of identifier.
- In this POC the identifier is an opening and closing HTML comment,
  since that won't affect what gets rendered in the viewport - though
there are problems with the way this POC adds them.
- In the CMS, when a form element gets updated, a message is sent to the
  preview iframe telling it what updated. Note that this POC only
touches input and textarea - select elements etc can be added too. The
POC also doesn't touch react which will be its own problem to tackle.
- Inside the iframe, when a message comes through from the CMS, it finds
  all comments for the updated field and updates the relevant elements.
- The POC ignores casting, so e.g. HTML from the WYSIWYG is
  encoded and MyDate.Nice will just render the raw date instead of the
nice date.

Check out the todo comments for caveats to this approach and shortcuts
I've taken.